requirements.txt
google-cloud-pubsub
pydantic
dependency-injector